flyttet amplitude instansen ut, grunnet at den må lazy-loades for at gatsby pre-compile ikke tuller seg
