+++ package.json
@@ -7,7 +7,6 @@
   "dependencies": {
     "@navikt/fnrvalidator": "^1.1.2",
     "@sanity/block-content-to-react": "^2.0.7",
-    "@types/amplitude-js": "^5.11.0",
     "@types/jest": "^26.0.3",
     "@typescript-eslint/parser": "^3.3.0",
     "amplitude-js": "^7.1.0",
@@ -63,6 +62,7 @@
     "typescript": "^3.9.5"
   },
   "devDependencies": {
+    "@types/amplitude-js": "^5.11.0",
     "@types/date-fns": "^2.6.0",
     "@types/node": "^14.0.13",
     "@types/react": "^16.9.38",
