+++ src/templates/faktaside/FaktaSide.tsx
@@ -11,7 +11,6 @@ import { FaktasideProvider } from './FaktaSideContext';
 import { FaktasideContext } from '../../../gatsby-utils/createFaktasider';
 import withErrorBoundary from '../../components/withErrorBoundary';
 import { useMount } from 'react-use';
-import { loggSidevisning } from '../../utils/logging';
 
 export interface FaktaSideProps extends PageProps<{}, FaktasideContext> {
   errors: any;
@@ -23,7 +22,7 @@ function FaktaSide(props: FaktaSideProps) {
   const erPublisert = page.visSprakversjon?.[lang];
   const tittel = page.title || '';
 
-  useMount(() => loggSidevisning(tittel));
+  useMount(() => import('../../utils/logging').then((logging) => logging.loggSidevisning(tittel)));
 
   if (!erPublisert) {
     return <IkkeOversatt {...props} />;
