+++ src/utils/logging.ts
@@ -1,27 +1,15 @@
-import amplitude from 'amplitude-js';
-import { isProduction } from './environment';
-
-const getApiKey = () => {
-  if (isProduction()) {
-    return 'edf391bf01b758a289ef5e7cb297f77a'; // prod
-  }
-  return '24eb6d83cfc9883c04c4eaec61251bf4'; // dev
-};
-
-const logging = amplitude.getInstance();
-logging.init(getApiKey(), '', {
-  apiEndpoint: 'amplitude.nav.no/collect',
-  saveEvents: false,
-  includeUtm: true,
-  batchEvents: false,
-  includeReferrer: true,
-});
+/*
+Vi må lazy-loade loggingConfig, da gatsby pre-rendrer html, og kresjer dermed amplitude-sdk
+Prøver derfor å ikke initiere amplitude-sdk instansen før en logevent blir kalt.
+*/
 
 const loggEvent = (event: string, ekstraData?: object) =>
-  logging.logEvent(event, {
-    ...ekstraData,
-    appName: 'dp-faktasider',
-  });
+  import('../utils/loggingConfig').then((logging) =>
+    logging.loggInstance.logEvent(event, {
+      ...ekstraData,
+      appName: 'dp-faktasider',
+    })
+  );
 
 export const loggError = (msg: string, ekstraData?: object) => loggEvent('Error', { ...ekstraData, msg });
